Separeted on/off method from SwitchbotDevice class
