+++ src/switchbot_utility/switchbot_device.py
@@ -25,15 +25,3 @@ class SwitchbotDevice(Switchbot):
         """Send command"""
         header = self.gen_sign()
         return requests.post("https://api.switch-bot.com/v1.1/devices/" + deviceId + "/commands", headers=header, data=json.dumps(body))
-
-    def turn_off(self):
-        """Turn off device"""
-        self._body['command'] = "turnOff"
-        result = self.command(self.deviceId, self._body)
-        return result.text
-
-    def turn_on(self):
-        """Turn on device"""
-        self._body['command'] = "turnOn"
-        result = self.command(self.deviceId, self._body)
-        return result.text
