+++ src/switchbot_utility/switchbot_ir_device.py
@@ -1,7 +1,8 @@
 from .switchbot_device import SwitchbotDevice
+from .onoff_ability import OnOffAbirity
 
 
-class SwitchbotIrDevice(SwitchbotDevice):
+class SwitchbotIrDevice(SwitchbotDevice, OnOffAbirity):
     """Switchbot virtual ir device"""
     def __init__(self, deviceId):
         """Constructor"""
